@@ -16,7 +16,6 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/cli": "^1.6.0",
     "@angular/common": "^5.1.0",
     "@angular/compiler": "^5.1.0",
     "@angular/compiler-cli": "^5.1.0",
@@ -26,7 +25,6 @@
     "@angular/platform-browser": "^5.1.0",
     "@angular/platform-browser-dynamic": "^5.1.0",
     "@angular/router": "^5.1.0",
-    "@ngtools/webpack": "^1.7.4",
     "angular-datatables": "^5.0.0",
     "angular2-image-upload": "^1.0.0-rc.0",
     "angular2-ladda": "^1.2.2",
@@ -42,18 +40,19 @@
     "rxjs": "^5.5.2",
     "tinymce": "^4.7.1",
     "ts-helpers": "^1.1.2",
-    "typescript": "^2.4.2",
+    "typescript": "^2.5.3",
     "zone.js": "^0.8.18"
   },
   "devDependencies": {
+    "@angular/cli": "^1.6.4",
     "@types/datatables.net": "^1.10.7",
-    "@types/jasmine": "2.8.2",
+    "@types/jasmine": "2.8.4",
     "@types/jquery": "^3.2.15",
-    "@types/node": "~8.0.47",
+    "@types/node": "~9.3.0",
     "codelyzer": "~4.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.3.0",
@@ -61,8 +60,8 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-remap-istanbul": "^0.6.0",
     "protractor": "~5.2.0",
-    "ts-node": "~3.3.0",
-    "tslint": "~5.8.0",
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
     "webdriver-manager": "12.0.6"
   }
 }
