@@ -1,7 +1,5 @@
-import {NgModule, ModuleWithProviders} from "@angular/core";
+import {ModuleWithProviders, NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {environment} from "../../environments/environment";
-import {BASE_PATH} from "../ws/soccer/variables";
 import {TranslationPipe} from "../pipes/translation.pipe";
 import {ValidationService} from "../services/validation.service";
 import {ErrorHandlerService} from "../services/error-handler.service";
@@ -9,12 +7,7 @@ import {TranslationService} from "../services/translation.service";
 import {LoginService} from "../services/login.service";
 import {AuthGuardAdminService} from "../services/auth-guard-admin.service";
 import {AuthGuardService} from "../services/auth-guard.service";
-import {AccountprofilerestcontrollerApi} from "../ws/soccer/api/AccountprofilerestcontrollerApi";
-import {AccountrestcontrollerApi} from "../ws/soccer/api/AccountrestcontrollerApi";
-import {RegistrationrestcontrollerApi} from "../ws/soccer/api/RegistrationrestcontrollerApi";
-import {AuthenticationcontrollerApi} from "../ws/soccer/api/AuthenticationcontrollerApi";
-import {HttpModule} from "@angular/http";
-import {ReactiveFormsModule, FormsModule} from "@angular/forms";
+import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {ReCaptchaModule} from "angular2-recaptcha";
 import {DataTablesModule} from "angular-datatables";
 import {SafeTranslationPipe} from "../pipes/safe-translation.pipe";
@@ -24,6 +17,7 @@ import {FocusOnSuccessDirective} from "../directives/focus-on-success.directive"
 import {FocusOnErrorDirective} from "../directives/focus-on-error.directive";
 import {RouterModule} from "@angular/router";
 import {PaginationComponent} from "../components/pagination/pagination.component";
+import {HttpClientModule} from "@angular/common/http";
 
 @NgModule({
     imports: [
@@ -32,7 +26,6 @@ import {PaginationComponent} from "../components/pagination/pagination.component
         ReCaptchaModule,
         FormsModule,
         ReactiveFormsModule,
-        HttpModule,
         RouterModule,
     ],
     declarations: [
@@ -49,7 +42,7 @@ import {PaginationComponent} from "../components/pagination/pagination.component
         ReCaptchaModule,
         FormsModule,
         ReactiveFormsModule,
-        HttpModule,
+        HttpClientModule,
         RouterModule,
         LoadingComponent,
         PaginationComponent,
@@ -60,24 +53,19 @@ import {PaginationComponent} from "../components/pagination/pagination.component
         FocusOnErrorDirective,
         CommonModule,
 
-    ],
+    ]
 })
 export class SharedModule {
     static forRoot(): ModuleWithProviders {
         return {
             ngModule: SharedModule,
             providers: [
-                {provide: BASE_PATH, useValue: environment.api_url},
-                AccountrestcontrollerApi,
-                AccountprofilerestcontrollerApi,
-                AuthenticationcontrollerApi,
-                RegistrationrestcontrollerApi,
                 AuthGuardService,
                 AuthGuardAdminService,
                 LoginService,
                 TranslationService,
                 ErrorHandlerService,
-                ValidationService
+                ValidationService,
             ]
         };
     }
