@@ -1,11 +1,9 @@
-import {Component, OnInit, Input, EventEmitter} from "@angular/core";
-import {FormGroup, FormBuilder, Validators} from "@angular/forms";
-import {NewsrestcontrollerApi} from "../../ws/soccer/api/NewsrestcontrollerApi";
+import {Component, EventEmitter, Input, OnInit} from "@angular/core";
+import {FormBuilder, FormGroup, Validators} from "@angular/forms";
 import {Router} from "@angular/router";
 import {ErrorHandlerService} from "../../services/error-handler.service";
-import {NewsDTO} from "../../ws/soccer/model/NewsDTO";
-import {SecUtil} from "../../classes/sec-util";
 import {ValidationService} from "../../services/validation.service";
+import {NewsDTO, NewsRestControllerService} from "../../ws/soccer";
 
 @Component({
     selector: 'app-edit-message-form',
@@ -55,7 +53,7 @@ export class EditMessageFormComponent implements OnInit {
         content: "",
     };
 
-    constructor(private _fb: FormBuilder, private _api: NewsrestcontrollerApi, private _router: Router, private _errorHandler: ErrorHandlerService, private _validationService: ValidationService) {
+    constructor(private _fb: FormBuilder, private _api: NewsRestControllerService, private _router: Router, private _errorHandler: ErrorHandlerService, private _validationService: ValidationService) {
     }
 
     ngOnInit() {
@@ -66,7 +64,8 @@ export class EditMessageFormComponent implements OnInit {
             id: ['', [<any>Validators.required]],
         });
 
-        this._api.getNews(this.messageId, SecUtil.getJwtHeaders()).subscribe(
+
+        this._api.getNews(this.messageId).subscribe(
             r => {
                 this.messageForm.patchValue({
                     content: r.content,
@@ -100,7 +99,8 @@ export class EditMessageFormComponent implements OnInit {
 
         if (isValid) {
             this.isLoading = true;
-            this._api.updateNews(model, SecUtil.getJwtHeaders()).subscribe(
+
+            this._api.updateNews(model).subscribe(
                 r => {
                     this._router.navigate(["/messages"]);
                 },
