@@ -1,9 +1,8 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {Routes, RouterModule} from "@angular/router";
+import {RouterModule, Routes} from "@angular/router";
 import {SharedModule} from "./shared.module";
 import {AlertModule} from "ngx-bootstrap";
-import {DoodlerestcontrollerApi} from "../ws/soccer/api/DoodlerestcontrollerApi";
 import {DoodleListComponent} from "../components/doodle-list/doodle-list.component";
 import {DoodleComponent} from "../components/doodle/doodle.component";
 import {LaddaModule} from "angular2-ladda";
@@ -33,9 +32,6 @@ const routes: Routes = [
         DoodleListComponent,
         DoodleComponent,
 
-    ],
-    providers: [
-        DoodlerestcontrollerApi,
     ]
 })
 export class DoodlesModule {
