@@ -1,9 +1,9 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {Routes, RouterModule} from "@angular/router";
+import {RouterModule, Routes} from "@angular/router";
 import {AuthGuardAdminService} from "../services/auth-guard-admin.service";
 import {SharedModule} from "./shared.module";
-import {AlertModule, BsDatepickerModule, DatepickerModule, TimepickerModule} from "ngx-bootstrap";
+import {AlertModule, BsDatepickerModule, TimepickerModule} from "ngx-bootstrap";
 import {EditMatchComponent} from "../components/edit-match/edit-match.component";
 import {CreateMatchComponent} from "../components/create-match/create-match.component";
 import {MatchesComponent} from "../components/matches/matches.component";
@@ -13,11 +13,10 @@ import {MatchComponent} from "../components/match/match.component";
 import {NextMatchComponent} from "../components/next-match/next-match.component";
 import {EditMatchFormComponent} from "../components/edit-match-form/edit-match-form.component";
 import {GoalsComponent} from "../components/goal/goal.component";
-import {SeasonsrestcontrollerApi} from "../ws/soccer/api/SeasonsrestcontrollerApi";
-import {MatchesrestcontrollerApi} from "../ws/soccer/api/MatchesrestcontrollerApi";
-import {ForecastrestcontrollerApi} from "../ws/weather/api/ForecastrestcontrollerApi";
-import {TeamsrestcontrollerApi} from "../ws/soccer/api/TeamsrestcontrollerApi";
 import {LaddaModule} from "angular2-ladda";
+import {environment} from "../../environments/environment";
+import {ForecastApiModule} from "../ws/weather/forecastApi.module";
+import {ForecastApiConfiguration} from "../ws/weather/forecastApiConfiguration";
 
 const routes: Routes = [
     {path: '', component: MatchesComponent},
@@ -34,6 +33,11 @@ const routes: Routes = [
         BsDatepickerModule.forRoot(),
         AlertModule.forRoot(),
         RouterModule.forChild(routes),
+        ForecastApiModule.forRoot(() => new ForecastApiConfiguration({
+            accessToken: environment.weather_api_key,
+            basePath: environment.weather_api_url,
+            country: environment.weather_api_country
+        })),
         LaddaModule.forRoot({
             style: "expand-right",
             spinnerSize: 20,
@@ -51,12 +55,6 @@ const routes: Routes = [
         EditMatchComponent,
         GoalsComponent,
         EditMatchFormComponent,
-    ],
-    providers: [
-        MatchesrestcontrollerApi,
-        SeasonsrestcontrollerApi,
-        ForecastrestcontrollerApi,
-        TeamsrestcontrollerApi,
     ]
 })
 export class MatchesModule {
