@@ -5,13 +5,11 @@ import {EditMessageComponent} from "../components/edit-message/edit-message.comp
 import {CreateMessageComponent} from "../components/create-message/create-message.component";
 import {TinymceComponent} from "../components/tinymce/tinymce.component";
 import {MessageFormComponent} from "../components/message-form/message-form.component";
-import {Routes, RouterModule} from "@angular/router";
+import {RouterModule, Routes} from "@angular/router";
 import {MessagesComponent} from "../components/messages/messages.component";
 import {AuthGuardAdminService} from "../services/auth-guard-admin.service";
 import {MessageItemComponent} from "../components/message-item/message-item.component";
 import {SharedModule} from "./shared.module";
-import {CommentsrestcontrollerApi} from "../ws/soccer/api/CommentsrestcontrollerApi";
-import {NewsrestcontrollerApi} from "../ws/soccer/api/NewsrestcontrollerApi";
 import {CommentFormComponent} from "../components/comment-form/comment-form.component";
 import {CommentComponent} from "../components/comment/comment.component";
 import {MessageComponent} from "../components/message/message.component";
@@ -51,10 +49,6 @@ const routes: Routes = [
         MessageItemComponent,
         CommentFormComponent,
         CommentComponent,
-    ],
-    providers: [
-        CommentsrestcontrollerApi,
-        NewsrestcontrollerApi,
     ]
 })
 export class MessageModule {
