@@ -1,11 +1,10 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
-import {Routes, RouterModule} from "@angular/router";
+import {RouterModule, Routes} from "@angular/router";
 import {SharedModule} from "./shared.module";
 import {AlertModule} from "ngx-bootstrap";
 import {MotmPollsComponent} from "../components/motm-polls/motm-polls.component";
 import {MotmPollComponent} from "../components/motm-poll/motm-poll.component";
-import {PollrestcontrollerApi} from "../ws/soccer/api/PollrestcontrollerApi";
 import {LaddaModule} from "angular2-ladda";
 
 const routes: Routes = [
@@ -29,9 +28,6 @@ const routes: Routes = [
     declarations: [
         MotmPollsComponent,
         MotmPollComponent,
-    ],
-    providers: [
-        PollrestcontrollerApi,
     ]
 })
 export class MotmModule {
