@@ -1,29 +1,30 @@
-export * from './AccountprofilerestcontrollerApi';
-import { AccountprofilerestcontrollerApi } from './AccountprofilerestcontrollerApi';
-export * from './AccountrestcontrollerApi';
-import { AccountrestcontrollerApi } from './AccountrestcontrollerApi';
-export * from './AuthenticationcontrollerApi';
-import { AuthenticationcontrollerApi } from './AuthenticationcontrollerApi';
-export * from './CommentsrestcontrollerApi';
-import { CommentsrestcontrollerApi } from './CommentsrestcontrollerApi';
-export * from './DoodlerestcontrollerApi';
-import { DoodlerestcontrollerApi } from './DoodlerestcontrollerApi';
-export * from './ErrorrestcontrollerApi';
-import { ErrorrestcontrollerApi } from './ErrorrestcontrollerApi';
-export * from './MatchesrestcontrollerApi';
-import { MatchesrestcontrollerApi } from './MatchesrestcontrollerApi';
-export * from './NewsrestcontrollerApi';
-import { NewsrestcontrollerApi } from './NewsrestcontrollerApi';
-export * from './PasswordrecoveryrestcontrollerApi';
-import { PasswordrecoveryrestcontrollerApi } from './PasswordrecoveryrestcontrollerApi';
-export * from './PollrestcontrollerApi';
-import { PollrestcontrollerApi } from './PollrestcontrollerApi';
-export * from './RegistrationrestcontrollerApi';
-import { RegistrationrestcontrollerApi } from './RegistrationrestcontrollerApi';
-export * from './SeasonsrestcontrollerApi';
-import { SeasonsrestcontrollerApi } from './SeasonsrestcontrollerApi';
-export * from './StatisticsrestcontrollerApi';
-import { StatisticsrestcontrollerApi } from './StatisticsrestcontrollerApi';
-export * from './TeamsrestcontrollerApi';
-import { TeamsrestcontrollerApi } from './TeamsrestcontrollerApi';
-export const APIS = [AccountprofilerestcontrollerApi, AccountrestcontrollerApi, AuthenticationcontrollerApi, CommentsrestcontrollerApi, DoodlerestcontrollerApi, ErrorrestcontrollerApi, MatchesrestcontrollerApi, NewsrestcontrollerApi, PasswordrecoveryrestcontrollerApi, PollrestcontrollerApi, RegistrationrestcontrollerApi, SeasonsrestcontrollerApi, StatisticsrestcontrollerApi, TeamsrestcontrollerApi];
+export * from './accountProfileRestController.service';
+import {AccountProfileRestControllerService} from './accountProfileRestController.service';
+import {AccountRestControllerService} from './accountRestController.service';
+import {AuthenticationControllerService} from './authenticationController.service';
+import {CommentsRestControllerService} from './commentsRestController.service';
+import {DoodleRestControllerService} from './doodleRestController.service';
+import {ErrorRestControllerService} from './errorRestController.service';
+import {MatchesRestControllerService} from './matchesRestController.service';
+import {NewsRestControllerService} from './newsRestController.service';
+import {PasswordRecoveryRestControllerService} from './passwordRecoveryRestController.service';
+import {PollRestControllerService} from './pollRestController.service';
+import {RegistrationRestControllerService} from './registrationRestController.service';
+import {SeasonsRestControllerService} from './seasonsRestController.service';
+import {StatisticsRestControllerService} from './statisticsRestController.service';
+import {TeamsRestControllerService} from './teamsRestController.service';
+
+export * from './accountRestController.service';
+export * from './authenticationController.service';
+export * from './commentsRestController.service';
+export * from './doodleRestController.service';
+export * from './errorRestController.service';
+export * from './matchesRestController.service';
+export * from './newsRestController.service';
+export * from './passwordRecoveryRestController.service';
+export * from './pollRestController.service';
+export * from './registrationRestController.service';
+export * from './seasonsRestController.service';
+export * from './statisticsRestController.service';
+export * from './teamsRestController.service';
+export const APIS = [AccountProfileRestControllerService, AccountRestControllerService, AuthenticationControllerService, CommentsRestControllerService, DoodleRestControllerService, ErrorRestControllerService, MatchesRestControllerService, NewsRestControllerService, PasswordRecoveryRestControllerService, PollRestControllerService, RegistrationRestControllerService, SeasonsRestControllerService, StatisticsRestControllerService, TeamsRestControllerService];
