@@ -1,29 +1,29 @@
-export * from './AccountDTO';
-export * from './AccountStatisticDTO';
-export * from './AddressDTO';
-export * from './AuthenticationRequestDTO';
-export * from './CommentDTO';
-export * from './DoodleDTO';
-export * from './ErrorDetailDTO';
-export * from './GoalDTO';
-export * from './ImageDTO';
-export * from './LocalizedMessageDTO';
-export * from './MatchDTO';
-export * from './MatchDoodleDTO';
-export * from './MatchPollDTO';
-export * from './MultipleChoiceVoteDTOlong';
-export * from './NewsDTO';
-export * from './PageDTOMatchDoodleDTO';
-export * from './PageDTOMatchPollDTO';
-export * from './PageDTONewsDTO';
-export * from './PasswordDTO';
-export * from './PasswordRecoveryDTO';
-export * from './PresenceDTO';
-export * from './ProfileDTO';
-export * from './RegistrationDTO';
-export * from './ResponseEntity';
-export * from './SeasonDTO';
-export * from './TeamDTO';
-export * from './ValidationErrorDTO';
-export * from './ValidationErrorDetailDTO';
-export * from './VotesDTO';
+export * from './accountDTO';
+export * from './accountStatisticDTO';
+export * from './addressDTO';
+export * from './authenticationRequestDTO';
+export * from './commentDTO';
+export * from './doodleDTO';
+export * from './errorDetailDTO';
+export * from './goalDTO';
+export * from './imageDTO';
+export * from './localizedMessageDTO';
+export * from './matchDTO';
+export * from './matchDoodleDTO';
+export * from './matchPollDTO';
+export * from './multipleChoiceVoteDTOlong';
+export * from './newsDTO';
+export * from './pageDTOMatchDoodleDTO';
+export * from './pageDTOMatchPollDTO';
+export * from './pageDTONewsDTO';
+export * from './passwordDTO';
+export * from './passwordRecoveryDTO';
+export * from './presenceDTO';
+export * from './profileDTO';
+export * from './registrationDTO';
+export * from './responseEntity';
+export * from './seasonDTO';
+export * from './teamDTO';
+export * from './validationErrorDTO';
+export * from './validationErrorDetailDTO';
+export * from './votesDTO';
